+++ src/components/Header/components/HeaderMobileNav.tsx
@@ -17,7 +17,7 @@ import { upperNavLinks, lowerNavLinks } from "../utils/headerUtils";
 /**
  * The Header navigation for the mobile view.
  */
-const MobileNav = chakra(() => {
+const HeaderMobileNav = chakra(() => {
   const styles = useMultiStyleConfig("HeaderMobileNav", {});
 
   return (
@@ -114,4 +114,4 @@ const MobileNav = chakra(() => {
   );
 });
 
-export default MobileNav;
+export default HeaderMobileNav;